made serializers and deserializers private
