+++ jackson-codec/src/main/org/bson/codecs/jackson/JacksonBsonDeserializer.java
@@ -1,4 +1,4 @@
-package org.bson.codecs.jackson.deserializers;
+package org.bson.codecs.jackson;
 
 import com.fasterxml.jackson.core.JsonGenerationException;
 import com.fasterxml.jackson.core.JsonParser;
@@ -11,7 +11,7 @@ import java.io.IOException;
 /**
  * Created by guo on 8/1/14.
  */
-public abstract class JacksonBsonDeserializer<T> extends JsonDeserializer<T> {
+abstract class JacksonBsonDeserializer<T> extends JsonDeserializer<T> {
     @Override
     public T deserialize(JsonParser jsonParser, DeserializationContext ctxt)
             throws IOException {
