+++ jackson-codec/src/main/org/bson/codecs/jackson/JacksonBsonDeserializers.java
@@ -1,4 +1,4 @@
-package org.bson.codecs.jackson.deserializers;
+package org.bson.codecs.jackson;
 
 import com.fasterxml.jackson.databind.module.SimpleDeserializers;
 
@@ -7,7 +7,7 @@ import java.util.Date;
 /**
  * Created by guo on 7/28/14.
  */
-public class JacksonBsonDeserializers extends SimpleDeserializers {
+class JacksonBsonDeserializers extends SimpleDeserializers {
     private static final long serialVersionUID = 261492073508623840L;
 
     /**
