+++ jackson-codec/src/main/org/bson/codecs/jackson/JacksonObjectIdSerializer.java
@@ -1,24 +1,26 @@
-package org.bson.codecs.jackson.serializers;
+package org.bson.codecs.jackson;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import org.bson.codecs.jackson.JacksonBsonGenerator;
+import org.bson.types.ObjectId;
 
 import java.io.IOException;
-import java.util.Date;
 
 /**
- * Created by guo on 7/30/14.
+ * Created by guo on 7/29/14.
  */
-public class JacksonDateSerializer extends JacksonBsonSerializer<Date> {
+public class JacksonObjectIdSerializer extends JacksonBsonSerializer<ObjectId> {
+
+
     @Override
-    public void serialize(Date value, JacksonBsonGenerator generator, SerializerProvider provider) throws IOException, JsonProcessingException {
+    public void serialize(ObjectId value, JacksonBsonGenerator generator, SerializerProvider provider) throws IOException, JsonProcessingException {
         if (value == null) {
             provider.defaultSerializeNull(generator);
         } else {
-            generator.writeDate(value);
+            generator.writeObjectId(value);
         }
     }
 }
