+++ jackson-codec/src/main/org/bson/codecs/jackson/JacksonSymbolSerializer.java
@@ -1,26 +1,26 @@
-package org.bson.codecs.jackson.serializers;
+package org.bson.codecs.jackson;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import org.bson.codecs.jackson.JacksonBsonGenerator;
-import org.bson.types.ObjectId;
+import org.bson.types.Symbol;
 
 import java.io.IOException;
 
 /**
- * Created by guo on 7/29/14.
+ * Created by guo on 7/30/14.
  */
-public class JacksonObjectIdSerializer extends JacksonBsonSerializer<ObjectId> {
-
+class JacksonSymbolSerializer extends JacksonBsonSerializer<Symbol> {
 
     @Override
-    public void serialize(ObjectId value, JacksonBsonGenerator generator, SerializerProvider provider) throws IOException, JsonProcessingException {
+    public void serialize(Symbol value, JacksonBsonGenerator generator, SerializerProvider provider) throws IOException, JsonProcessingException {
         if (value == null) {
             provider.defaultSerializeNull(generator);
         } else {
-            generator.writeObjectId(value);
+            generator.writeSymbol(value);
         }
     }
 }
+
